+++ src/index.js
@@ -1,14 +1,14 @@
-export { default as Text } from './Text'
 export { default as Box } from './Box'
-export { default as Flex } from './Flex'
-export { default as Image } from './Image'
 export { default as BackgroundImage } from './BackgroundImage'
-export { default as Button } from './Button/Button'
-export { default as GreenButton } from './Button/GreenButton'
-export { default as RedButton } from './Button/RedButton'
-export { default as OutlineButton } from './Button/OutlineButton'
+export { default as Button } from './Button'
+export { default as GreenButton } from './GreenButton'
+export { default as RedButton } from './RedButton'
+export { default as OutlineButton } from './OutlineButton'
+export { default as Flex } from './Flex'
 export { default as Heading } from './Heading'
 export { default as Icon } from './Icon'
+export { default as Image } from './Image'
+export { default as Text } from './Text'
 
 export { default as theme } from './theme'
 export { default as legacyTheme } from './legacyTheme'
